requirements.txt
# Core AI Framework
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-openai>=0.2.0
anthropic>=0.34.0
openai>=1.50.0

# MCP Integration
mcp>=1.25.0
httpx==0.25.2
anyio>=4.12.0

# Google APIs
google-api-python-client==2.120.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-cloud-storage==2.14.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
tenacity==8.2.3
pytz==2024.1

# Logging & Monitoring
structlog==24.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.23.0
black==23.12.1
mypy==1.8.0

